Refactor and clean import to extract -> glean pkg
